pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "glue-pyspark-dev-tools"
repository = "https://github.com/dashmug/glue-pyspark-dev-tools"
version = "0.1.0"
package-mode = false
description = "Glue PySpark Development Tools"
authors = ["Noel Llevares <dashmug@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10,<3.11"
pip = "^24.0"

[tool.poetry.group.runtime.dependencies]
# The "runtime" dependency group refers to dependencies that already
# exist in AWS Glue's runtime. We don't need to install these in the
# container.
#
# It is best to keep the dependencies here in sync with what's on Glue's
# documentation.
# https://docs.aws.amazon.com/glue/latest/dg/aws-glue-programming-python-libraries.html#glue-modules-provided
aws-glue-libs = { git = "https://github.com/awslabs/aws-glue-libs.git", rev = "master" }
pyspark = "3.3.0"

[tool.poetry.group.dev.dependencies]
pytest-randomly = "^3.15.0"
pytest-cov = "^4.1.0"
pip-audit = "^2.7.2"
mypy = "^1.9.0"
ruff = "^0.3.2"
pytest = "^8.1.1"
pre-commit = "^3.6.2"
import-linter = "^2.0"


[tool.poetry.group.deploy.dependencies]
aws-cdk-lib = "^2.133.0"
cdk-nag = "^2.28.65"

[tool.ruff]
line-length = 88
output-format = "full"
respect-gitignore = true
target-version = "py310"
exclude = ["snapshots", ".venv", ".ipynb_checkpoints"]

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    # Line too long
    "E501",
    # pydocstyle
    "D",
    # Missing type annotation for `self` in method
    "ANN101",
    # To Do comments
    "TD002",
    "TD003",
    "FIX002",
    # Conflicts with ruff format
    "COM812",
    "ISC001",
]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.lint.mccabe]
max-complexity = 7

[tool.ruff.lint.per-file-ignores]
"**/test_*" = ["S101", "ANN", "PT", "PD"]

[tool.mypy]
files = "**/*.py"
python_version = "3.10"
show_error_codes = true
pretty = true
strict = true
ignore_missing_imports = true
implicit_reexport = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = "*.tests.*"
disallow_incomplete_defs = false
disallow_untyped_defs = false
disallow_untyped_calls = false

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::FutureWarning:.*pyspark.*",
]

[tool.coverage.run]
omit = ["**/test_*.py"]

[tool.coverage.report]
skip_empty = true
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.importlinter]
root_packages = ["infrastructure", "jobs"]
include_external_packages = true

[[tool.importlinter.contracts]]
name = "Jobs and Infrastructure code should not import from each other"
type = "independence"
modules = ["jobs", "infrastructure"]

[[tool.importlinter.contracts]]
name = "Glue Job scripts should not use AWS CDK package"
type = "forbidden"
source_modules = ["jobs"]
forbidden_modules = ["aws_cdk"]

[[tool.importlinter.contracts]]
name = "Infrastructure code should follow layers architecture"
type = "layers"
layers = ["infrastructure.stacks", "infrastructure.constructs"]
